+++ str/json_test.go
@@ -1,0 +1,9 @@
+package str
+
+import "testing"
+
+func TestJSON(t *testing.T) {
+	// TODO: JSON
+	// TODO: IndentedJSON
+	// TODO: MarshalJSON
+}