+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * conf.LoadEnv
   * dict.Inspect
   * Sys package and display functions
+  * String JSON functions
   * Root package
 ## v0.3.6 - Schema Combos 
   * Commit: 2026-04-01 22:44